+++ source/Time.cpp
@@ -3,7 +3,7 @@
 std::chrono::time_point<std::chrono::high_resolution_clock> Time::loopStartTime;
 std::chrono::time_point<std::chrono::high_resolution_clock> Time::loopEndTime;
 std::chrono::time_point<std::chrono::high_resolution_clock> Time::initTime;
-double Time::deltaTime = 0.0;
+float Time::deltaTime = 0.0;
 
 void Time::Initialize()
 {
@@ -13,24 +13,24 @@ void Time::Initialize()
 void Time::CalculateLoopTime()
 {
     loopEndTime = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double> delta = loopEndTime - loopStartTime;
+    std::chrono::duration<float> delta = loopEndTime - loopStartTime;
     deltaTime = delta.count();
     loopStartTime = loopEndTime;  // Prepare for the next loop
 }
 
-double Time::DeltaTime()
+float Time::DeltaTime()
 {
     return deltaTime;
 }
 
-double Time::DoubleDeltaTime()
+float Time::DoubleDeltaTime()
 {
-    return 2.0 * deltaTime;
+    return 2.0f * deltaTime;
 }
 
-double Time::TimeSinceInit()
+float Time::TimeSinceInit()
 {
     auto now = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double> elapsed = now - initTime;
+    std::chrono::duration<float> elapsed = now - initTime;
     return elapsed.count();
 }
