+++ source/Time.h
@@ -8,13 +8,13 @@ private:
     static std::chrono::time_point<std::chrono::high_resolution_clock> loopStartTime;
     static std::chrono::time_point<std::chrono::high_resolution_clock> loopEndTime;
     static std::chrono::time_point<std::chrono::high_resolution_clock> initTime;
-    static double deltaTime;
+    static float deltaTime;
 
 
 public:
     static void Initialize();
     static void CalculateLoopTime();
-    static double DeltaTime();
-    static double DoubleDeltaTime();
-    static double TimeSinceInit();
+    static float DeltaTime();
+    static float DoubleDeltaTime();
+    static float TimeSinceInit();
 };
